# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HermesX API
  description: Agent-first Runtime Control Plane for governed, multi-tenant AI automation.
  version: 2.4.0-dev
  contact:
    name: HermesX Team
    url: https://github.com/Colin4k1024/hermesx
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.hermesx.dev
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /v1/sessions:
    get:
      tags: [Sessions]
      summary: List user sessions
      operationId: listSessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Sessions]
      summary: Create a new session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/memories:
    get:
      tags: [Memory]
      summary: List memories
      operationId: listMemories
      parameters:
        - name: X-Hermes-User-Id
          in: header
          schema:
            type: string
          description: User ID to query (admin only)
      responses:
        '200':
          description: List of memories
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant_id:
                    type: string
                  user_id:
                    type: string
                  memories:
                    type: array
                    items:
                      $ref: '#/components/schemas/MemoryEntry'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key in format hk_...
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
        tenant_id:
          type: string
        user_id:
          type: string
        model:
          type: string
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
    SessionCreateRequest:
      type: object
      properties:
        model:
          type: string
        user_id:
          type: string
...
    MemoryEntry:
      type: object
      properties:
        key:
          type: string
        content:
          type: string
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: authorization required
...